Create an HelpRepository
The help messages are currently located in `Help\ Documentation/Commands`. Although it works, the help messages are part of the source code and should be bundled as resources in the jar file.

This commit create a repository that can fetch these help messages as resources

Although not used in the code, everything is set in place to do so